+++ src/analyzers/resistor-color/index.ts
@@ -1,323 +1,262 @@
-import { AST_NODE_TYPES } from "@typescript-eslint/typescript-estree";
-import { Program } from "@typescript-eslint/typescript-estree/dist/ts-estree/ts-estree";
+import { AST_NODE_TYPES, TSESTree } from "@typescript-eslint/typescript-estree";
 
-import { isReturnBlockStatement } from "~src/analyzers//utils/is_return_block_statement";
-import { isReturnStatementWithValue } from "~src/analyzers//utils/is_return_statement_with_value";
-import { AnalyzerImpl } from "~src/analyzers/AnalyzerImpl";
-import { extractExport } from "~src/analyzers/utils/extract_export";
-import { extractMainMethod, MainMethod } from "~src/analyzers/utils/extract_main_method";
-import { parameterName } from '~src/analyzers/utils/extract_parameter';
-import { findTopLevelConstants } from "~src/analyzers/utils/find_top_level_constants";
-import { isCallExpression } from "~src/analyzers/utils/is_call_expression";
-import { isIdentifier } from "~src/analyzers/utils/is_identifier";
-import { annotateType } from "~src/analyzers/utils/type_annotations";
 import { factory } from "~src/comments/comment";
-import { NO_METHOD, NO_NAMED_EXPORT, NO_PARAMETER, UNEXPECTED_SPLAT_ARGS, PARSE_ERROR } from "~src/comments/shared";
-import { AstParser, ParsedSource } from "~src/parsers/AstParser";
-import { NoSourceError } from "~src/errors/NoSourceError";
-import { ParserError } from "~src/errors/ParserError";
+import { NO_METHOD, NO_NAMED_EXPORT, NO_PARAMETER, BETA_COMMENTARY_PREFIX, UNEXPECTED_PARAMETER } from "~src/comments/shared";
+import { NoExportError } from "~src/errors/NoExportError";
+import { NoMethodError } from "~src/errors/NoMethodError";
+import { AstParser } from "~src/parsers/AstParser";
 
-/**
- * The factories here SHOULD be kept in sync with exercism/website-copy. Under
- * normal use, they do NOT dictate the actual commentary output of the analyzer,
- * as that is provided by the website-copy repo.
- *
- * https://github.com/exercism/website-copy/tree/master/automated-comments/javascript/resistor-color
- */
+import { IsolatedAnalyzerImpl } from "../IsolatedAnalyzerImpl";
+import { ResistorColorSolution, HelperNotOptimal, HelperCallNotFound, MethodNotFound, MissingExpectedCall } from "./ResistorColorSolution";
 
-const TIP_EXPORT_INLINE = factory`
+const TIP_EXPORT_INLINE = factory<'method.signature' | 'constant.signature'>`
 Did you know that you can export functions, classes and constants directly
 inline?
+\`\`\`javascript
+export ${'constant.signature'}
+
+export ${'method.signature'}
+\`\`\`
 `('javascript.resistor-color.export_inline')
+
+const SIGNATURE_NOT_OPTIMAL = factory`
+📕 If you look at the tests, the function \`colorCode\` only receives one
+parameter. Nothing more and nothing less.
+
+📕 Remove the additional parameters from your function, as their value will
+always be \`undefined\` or whatever default you've assigned.
+`('javascript.resistor-color.signature_not_optimal')
+
+const USE_INDEX_OF = factory<'current'>`
+💬 Replace \`${'current'}\` with a different built-in function, a function
+which does exactly what is now explicitly programmed in: finding the index
+of a given value in an \`Array\`.
+`('javascript.resistor-color.use_index_of')
+
+const USE_IMPLICIT_INCLUDES = factory<'current'>`
+💬 Remove the explicit existence check \`.includes\`. When the color-code
+is received, there is a special value that is returned if the color is not
+present. Use that special value instead.
+`('javascript.resistor-color.use_implicit_includes')
+
+const DONT_NORMALISE_INPUTS = factory`
+💬 Remove the call to \`.toLowerCase\`. The tests only provide the inputs
+in lower case, and the colors should be defined in lower case. There is
+no need to manually normalise the inputs.
+`('javascript.resistor-color.dont_normalise_inputs')
+
+const USE_ARRAY_COMPREHENSIONS = factory<'current'>`
+💬 Replace \`${'current'}\` with a single function call that _directly_
+finds the index of a given input value.
+`('javascript.resistor-color.use_array_comprehensions')
+
+const ISSUE_OPTIMISE_HELPER = factory<'method.name'>`
+⚡ The helper method \`${'method.name'}\` is not optimal. The helper can
+probably be the same as the solution to \`resistor-color\`. Mentor the student
+to retrieve their solution and/or optimise their helper.
+`('javascript.resistor-color.must_optimise_helper')
+
+const ISSUE_USE_A_HELPER = factory`
+📕 Mentor the student to add helper function and DRY-up this solution. The
+solution to \`resistor-color\` can be used as helper method here. When using an
+\`Array\` as colors source, in a years time, will the student recall why it's
+the _index_ in that array? When using an \`Object\`, what does the value mean?
+Re-using \`colorCode\` explains this in both cases.
+
+💬 Using a helper method is good practice, because it replaces a cryptic "member
+call" with a named call, that can be documented individually.
+`('javascript.resistor-color.must_use_a_helper')
+
+const ISSUE_METHOD_NOT_FOUND = factory<'method.name'>`
+⚡ Ensure the method \`${'method/name'}\` exists. It was not found when
+analysing this solution. If it does not exist, point this out to the student.
+
+`('javascript.resistor-color.must_declare_function')
+
+const ISSUE_EXPECTED_CALL = factory<'method.name' | 'expected.reason'>`
+📕 In order to ${'expected.reason'}, expected a \`${'method.name'}\` call. If
+that reasoning applies, mentor the student to add this call.
+`('javascript.resistor-color.must_add_missing_call')
+
+type Program = TSESTree.Program
 
 const Parser: AstParser = new AstParser(undefined, 1)
 
-const NOT_FOUND = {} as const
+export class ResistorColorAnalyzer extends IsolatedAnalyzerImpl {
 
-export class ResistorColorAnalyzer extends AnalyzerImpl {
-
-  private program!: Program
-  private source!: string
-
-  private _mainMethod!: ReturnType<typeof extractMainMethod>
-  private _mainConstant!: ReturnType<typeof findTopLevelConstants>[0] | typeof NOT_FOUND
-  private _mainExports!: { function: ReturnType<typeof extractExport>; constant: ReturnType<typeof extractExport> }
-
-  private get mainMethod(): ReturnType<typeof extractMainMethod> {
-    if (!this._mainMethod) {
-      this._mainMethod = extractMainMethod(this.program, 'colorCode')
-    }
-    return this._mainMethod
-  }
-
-  private get mainExports(): { function: ReturnType<typeof extractExport>; constant: ReturnType<typeof extractExport> } {
-    if (!this._mainExports) {
-      this._mainExports = {
-        function: extractExport(this.program, 'colorCode'),
-        constant: extractExport(this.program, 'COLORS')
-      }
-    }
-    return this._mainExports
-  }
-
-  private get mainConstant(): ReturnType<typeof findTopLevelConstants>[0] | typeof NOT_FOUND {
-    if (!this._mainConstant) {
-      this._mainConstant = findTopLevelConstants(this.program, ['let', 'const', 'var']).find(
-        (declaration): boolean => isIdentifier(declaration.id, 'COLORS')
-      ) || NOT_FOUND
-    }
-
-    return this._mainConstant
-  }
-
-  public async execute(input: Input): Promise<void> {
-    const [parsed] = await this.parse(input)
-
-    this.program = parsed.program
-    this.source = parsed.source
+  protected async execute(input: Input, output: WritableOutput): Promise<void> {
+    const [parsed] = await Parser.parse(input)
 
     // Firstly we want to check that the structure of this solution is correct
     // and that there is nothing structural stopping it from passing the tests
-    this.checkStructure()
+    const solution = this.checkStructure(parsed.program, parsed.source, output)
 
     // Now we want to ensure that the method signature is sane and that it has
-    // valid arguments
-    this.checkSignature()
+    // valid arguments.
+    this.checkSignature(solution, output)
 
-    // There are a handful optimal solutions for resistor-color which needs no
+    // There are a handful optimal solutions for gigasecond which needs no
     // comments and can just be approved. If we have it, then let's just
     // acknowledge it and get out of here.
-    this.checkForOptimalSolutions()
+    this.checkForOptimalSolutions(solution, output)
+
+    // The solution might not be optimal but still be approvable. Check these
+    // first and bail-out (with approval) if that's the case.
+    this.checkForApprovableSolutions(solution, output)
+
+    // Time to find sub-optimal code.
+    this.checkForDisapprovables(solution, output)
 
     // The solution is automatically referred to the mentor if it reaches this
   }
 
-  private async parse(input: Input): never | Promise<ParsedSource[]> {
+  private checkStructure(program: Readonly<Program>, source: Readonly<string>, output: WritableOutput): ResistorColorSolution | never {
     try {
-      return await Parser.parse(input)
-    } catch (err) {
-      if (err instanceof NoSourceError) {
-        this.logger.error(`=> [NoSourceError] ${err.message}`)
-        this.redirect()
+      return new ResistorColorSolution(program, source)
+    } catch (error) {
+      if (error instanceof NoMethodError) {
+        output.disapprove(NO_METHOD({ 'method.name': error.method }))
       }
 
-      if (err instanceof ParserError) {
-        this.logger.error(`=> [ParserError] ${err.message}`)
-        const { message, ...details } = err.original
-        this.disapprove(PARSE_ERROR({ error: message, details: JSON.stringify(details) }))
+      if (error instanceof NoExportError) {
+        output.disapprove(NO_NAMED_EXPORT({ 'export.name': error.namedExport }))
       }
 
-      throw err
+      throw error
     }
   }
 
-  private checkStructure(): void | never {
-    const method = this.mainMethod
-    const { function: [functionDeclaration,], constant: [constantDeclaration, ] } = this.mainExports
-
-    // First we check that there is a two-fer function and that this function
-    // is exported.
-    if (!method) {
-      this.comment(NO_METHOD({ 'method.name': 'colorCode' }))
+  private checkSignature({ entry }: ResistorColorSolution, output: WritableOutput): void | never {
+    // If there is no parameter then this solution won't pass the tests.
+    //
+    if (!entry.hasAtLeastOneParameter) {
+      output.disapprove(NO_PARAMETER({ 'function.name': entry.name }))
     }
 
-    if (!functionDeclaration) {
-      this.comment(NO_NAMED_EXPORT({ 'export.name': 'colorCode' }))
+    // If this is not a simple parameter, but something else such as a splat,
+    // or a parameter with a default argument, bail out and refer to mentor.
+    //
+    if (!entry.hasSimpleParameter) {
+      output.redirect(UNEXPECTED_PARAMETER({ type: entry.parameterType }))
     }
 
-    if (!constantDeclaration) {
-      this.comment(NO_NAMED_EXPORT({ 'export.name': 'COLORS' }))
+    // If there is more than one parameter, something fishy is going on. Collect
+    // the comment for the student, to disapprove later on.
+    //
+    if (!entry.hasExactlyOneParameter) {
+      output.add(SIGNATURE_NOT_OPTIMAL())
     }
 
-    if (this.hasCommentary) {
-      this.disapprove()
-    }
-  }
+    // TODO: do we want to check more here?
 
-  private checkSignature(): void | never {
-    const method: MainMethod = this.mainMethod!
-
-    // If there is no parameter
-    // then this solution won't pass the tests.
-    if (method.params.length === 0) {
-      this.disapprove(NO_PARAMETER({ 'function.name': method.id!.name }))
-    }
-
-    const firstParameter = method.params[0]
-
-    // If they provide a splat, the tests can pass but we should suggest they
-    // use a real parameter.
-    if (firstParameter.type === AST_NODE_TYPES.RestElement) {
-      const splatArgName = parameterName(firstParameter)
-      const splatArgType = annotateType(firstParameter.typeAnnotation)
-
-      this.disapprove(UNEXPECTED_SPLAT_ARGS({
-        'splat-arg.name': splatArgName,
-        'parameter.type': splatArgType
-      }))
+    //
+    //
+    if (output.hasCommentary) {
+      output.disapprove()
     }
   }
 
-  private checkForOptimalSolutions(): void | never {
+  private checkForOptimalSolutions(solution: ResistorColorSolution, output: WritableOutput): void | never {
     // The optional solution looks like this:
     //
-    // export const COLORS = ['...', '...']
-    // export function colorCode(color) {
-    //   return COLORS.indexOf(color)
+    // const COLORS = [...]
+    //
+    // function colorCode(...) { ... }
+    //
+    // export function value([tens, ones]) {
+    //   return colorCode(ones) + colorCode(tens) * 10
     // }
     //
-    // The COLORS constant must be an Array and there must be a single call to
-    // indexOf. Additionally, the function can not have a default argument, or
-    // any other syntax, such as conditionals, re-assignment, and so forth.
-    //
-    // Other solutions might be approved but this
-    // is the only one that we would approve without comment.
-    //
 
-    if (
-    // !this.isArgumentOptimal()
-      !this.isOneCallSolution()
-      // || !this.isUsingArrayColors()
-      // || !this.isUsingIndexOf()
-    ) {
+    if (!solution.isOptimal()) {
       // continue analyzing
-      this.logger.log('~> Solution is not optimal')
+      this.logger.log('~> solution is not optimal')
+      this.processLastIssue(solution, output)
       return
     }
 
-    this.checkForTips()
-    this.approve()
+    this.checkForTips(solution, output)
+    output.approve()
   }
 
-  private checkForTips(): void | never {
-    if (!this.hasInlineExport()) {
-      this.comment(TIP_EXPORT_INLINE())
+  private processLastIssue(solution: ResistorColorSolution, output: WritableOutput): void | never {
+    const lastIssue = solution.entry.lastIssue
+    if (!lastIssue) {
+      this.logger.log('~> no entry issue found')
+      return
+    }
+
+    if (lastIssue instanceof HelperNotOptimal) {
+      // output.add(BETA_COMMENTARY_PREFIX())
+      output.disapprove(ISSUE_OPTIMISE_HELPER({ 'method.name': lastIssue.helperName }))
+    } else if (lastIssue instanceof HelperCallNotFound) {
+      // output.add(BETA_COMMENTARY_PREFIX())
+      output.disapprove(ISSUE_USE_A_HELPER())
+    } else if (lastIssue instanceof MethodNotFound) {
+      // output.add(BETA_COMMENTARY_PREFIX())
+      output.disapprove(ISSUE_METHOD_NOT_FOUND({ 'method.name': lastIssue.methodName }))
+    } else if (lastIssue instanceof MissingExpectedCall) {
+      // output.add(BETA_COMMENTARY_PREFIX())
+      output.add(ISSUE_EXPECTED_CALL({ 'method.name': lastIssue.methodName, 'expected.reason': lastIssue.reason }))
+
+      output.disapprove()
+    } else {
+      this.logger.error('The analyzer did not handle the issue: ' + JSON.stringify(lastIssue))
+      output.redirect()
     }
   }
 
-  private isOneCallSolution(): boolean | never {
-    // Maximum body count may be 3 (3 - 1 + 1)
-    //
-    // 1: export function colorCode(color) {
-    // 2:  return ...
-    // 3: }
-    //
-    // but can also be less 1 (1 - 1 + 1)
-    //
-    // 1: export const colorCode = (color) => ...
-    //
-    const method: MainMethod = this.mainMethod!
-    const body = method.body!
+  private checkForApprovableSolutions(solution: ResistorColorSolution, output: WritableOutput): void | never {
+    if (solution || output) {
+      return
+    }
+  }
 
-    // This trick actually looks to the inner exppresion instead of the entire
-    // function in order to allow for comments inside the body.
-    const { loc: { start: { line: lineStart }, end: { line: lineEnd } } } =
-      isReturnBlockStatement(body)
-        ? body.body[0]
-        : method
+  private checkForDisapprovables(solution: ResistorColorSolution, output: WritableOutput): void | never {
+    const numberOfComments = output.comments.length
 
-    this.logger.log(`=> Body consists of ${lineEnd - lineStart + 1} lines and is a ${method.type}.`)
-
-    if ((lineEnd - lineStart + 1) > 3) {
-      return false
+    if (solution.entry.hasFindIndex) {
+      output.add(USE_INDEX_OF({ current: '.findIndex(...)' }))
+    } else if (solution.entry.hasForEach) {
+      output.add(USE_ARRAY_COMPREHENSIONS({ current: '.forEach(...)' }))
+    } else if (solution.entry.hasFor) {
+      output.add(USE_ARRAY_COMPREHENSIONS({ current: 'for(...) { }' }))
     }
 
-    // There can only be a single call to COLORS.indexOf and there may not be
-    // an other tokens, whatsoever. The following checks are only testing the
-    // optimal variations.
+    if (solution.entry.hasIncludes) {
+      output.add(USE_IMPLICIT_INCLUDES())
+    }
 
-    // function colorCode(color) { return COLORS.indexOf(color) }
-    //
-    if (method.type === AST_NODE_TYPES.FunctionDeclaration) {
-      this.logger.log(`=> The Function Declaration has ${method.params.length} argument(s)`)
+    if (solution.entry.hasToLowerCase) {
+      output.add(DONT_NORMALISE_INPUTS())
+    }
 
-      if (method.params.length !== 1) {
-        return false
+    if (numberOfComments < output.commentCount) {
+      // output.comments.unshift(BETA_COMMENTARY_PREFIX())
+      output.disapprove()
+    }
+
+    if (solution || output) {
+      return
+    }
+  }
+
+  private checkForTips(solution: ResistorColorSolution, output: WritableOutput): void | never {
+    if (!solution.hasInlineExports) {
+      if (!output.hasCommentary) {
+        // output.add(BETA_COMMENTARY_PREFIX())
       }
 
-      const functionParameter = method.params[0]
-      const callExpression = method.body
-        && isReturnBlockStatement(method.body)
-        && isReturnStatementWithValue(method.body.body[0])
-        && isCallExpression(method.body.body[0].argument, 'COLORS', 'indexOf')
-        && method.body.body[0].argument
-
-      return !!(
-        isIdentifier(functionParameter)
-        && callExpression
-        && isIdentifier(callExpression.arguments[0], functionParameter.name)
+      // export { gigasecond }
+      output.add(
+        TIP_EXPORT_INLINE({
+          'method.signature': solution.entry.signature,
+          'constant.signature': solution.constant.signature
+        })
       )
     }
 
-    // const colorCode = (color) => COLORS.indexOf(color)
-    // const colorCode = (color) => { return COLORS.indexOf(color) }
-    //
-    if (method.type === AST_NODE_TYPES.ArrowFunctionExpression) {
-      this.logger.log(`=> The Arrow Function Expression has ${method.params.length} argument(s)`)
-
-      if (method.params.length !== 1) {
-        return false
-      }
-
-      const functionParameter = method.params[0]
-
-      // First test/option is the implicit return
-      // Second test/option is the explicit return
-      //
-      const callExpression = (isCallExpression(body, 'COLORS', 'indexOf') && body)
-        || (
-          isReturnBlockStatement(body)
-          && isReturnStatementWithValue(body.body[0])
-          && isCallExpression(body.body[0].argument, 'COLORS', 'indexOf')
-          && body.body[0].argument
-        )
-
-      return isIdentifier(functionParameter)
-        && callExpression
-        && isIdentifier(callExpression.arguments[0], functionParameter.name)
-    }
-
-    // const colorCode = function (color) { return COLORS.indexOf(color) }
-    //
-    if (method.type === AST_NODE_TYPES.FunctionExpression) {
-      this.logger.log(`=> The Function Expression has ${method.params.length} argument(s)`)
-
-      if (method.params.length > 1) {
-        return false
-      }
-
-      const functionParameter = method.params[0]
-      const callExpression = body
-        && isReturnBlockStatement(body)
-        && isReturnStatementWithValue(body.body[0])
-        && isCallExpression(body.body[0].argument, 'COLORS', 'indexOf')
-        && body.body[0].argument
-
-      return isIdentifier(functionParameter)
-        && callExpression
-        && isIdentifier(callExpression.arguments[0], functionParameter.name)
-    }
-
-    // Should _never_ happen
-    this.logger.log(`=> The body failed all the stuctural tests. It's a ${method!.type}.`)
-    // Bail out
-    return this.redirect()
-  }
-
-  private hasInlineExport(): boolean  {
-    // Additionally make sure the export is inline by checking if it doesn't
-    // have any specifiers:
-    //
-    // export function value
-    // => no specififers
-    //
-    // export { value }
-    // => yes specififers
-    //
-    return this.mainExports.function[0]!.specifiers
-      && this.mainExports.function[0]!.specifiers.length === 0
-      && this.mainExports.constant[0]!.specifiers
-      && this.mainExports.constant[0]!.specifiers.length === 0
+    // TODO optimize param
+    // TODO use helper method
   }
 }
-