+++ test/analyzers/resistor-color/smoke.ts
@@ -7,10 +7,14 @@ describe('When running analysis on resistor-color', () => {
   it('can approve as optimal', async () => {
 
     const solutionContent = `
-    export const COLORS = ['...', '...']
+    export const COLORS = [
+      "black", "brown", "red", "orange", "yellow",
+      "green", "blue", "violet", "grey", "white"
+    ]
     export function colorCode(color) {
       return COLORS.indexOf(color)
     }
+
     `.trim()
 
     const output = await analyze(solutionContent)